"Deprecate" old UIKeyConfig and start from clean slate
